+++ pkgs/objective_c/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 7.2.0-wip
 
 - Use ffigen 18.2.0
-- NSArray is now a Dart Iterable and NSMutableArray is now a Dart List.
+- `NSArray` is now a Dart `Iterable` and `NSMutableArray` is now a Dart `List`.
+- Add `.toNSNumber()` extension method to `int`, `double`, and `num`.
 
 ## 7.1.0